+++ test/packages/monorepo/npm/lib/camel/src/camel.js
@@ -1,0 +1,5 @@
+"use strict";
+
+const camelcase = require("camelcase");
+
+module.exports = (val) => camelcase(val);