+++ test/script.js
@@ -38,6 +38,7 @@ const CONFIGS = [
 const SCENARIOS = [
   "simple",
   "individually",
+  "monorepo",
   "webpack",
   "huge"
 ]
@@ -48,6 +49,11 @@ const TIMING_SCENARIOS = new Set([
   "simple",
   "individually",
   "huge"
+]);
+
+// Some scenarios are only feasible in Jetpack
+const JETPACK_ONLY_SCENARIOS = new Set([
+  "monorepo"
 ]);
 
 const MATRIX = SCENARIOS
@@ -72,14 +78,23 @@ const h2 = (msg) => log(chalk `\n{cyan ## ${msg}}`);
 const build = async () => {
   const clean = [
     "**",
+    "!functions/*/node_modules/**",
+    "!functions/*/package-lock.json",
+    "!lib/*/node_modules/**",
+    "!lib/*/package-lock.json",
     "!node_modules/**",
-    "!package-lock.json"
+    "!package-lock.json",
+    "!lerna.json"
   ];
   const patterns = [
     "package.json",
     "serverless.*",
     "src/**",
-    "*.js"
+    "*.js",
+    "functions/*/src/**",
+    "functions/*/package.json",
+    "lib/*/src/**",
+    "lib/*/package.json"
   ];
 
   for (const scenario of SCENARIOS) {
@@ -188,29 +203,37 @@ const benchmark = async () => {
           path.join(pluginArchive, path.basename(zipFile))
         )));
 
-        logTask("[task:start:baseline]");
-        const baselineTime = await runPackage();
-        logTask("[task:end:baseline]");
+        let baselineTime;
+        if (JETPACK_ONLY_SCENARIOS.has(scenario)) {
+          logTask("[task:skipping:baseline]");
+        } else {
+          logTask("[task:start:baseline]");
+          baselineTime = await runPackage();
+          logTask("[task:end:baseline]");
 
-        const baselineArchive = path.join(archiveRoot, scenario, mode, "baseline");
-        await del(baselineArchive);
-        await fs.mkdirp(baselineArchive);
-        const baselineZips = await globby(".serverless/*.zip", { cwd });
-        await Promise.all(baselineZips.map((zipFile) => fs.copy(
-          path.join(cwd, zipFile),
-          path.join(baselineArchive, path.basename(zipFile))
-        )));
+          const baselineArchive = path.join(archiveRoot, scenario, mode, "baseline");
+          await del(baselineArchive);
+          await fs.mkdirp(baselineArchive);
+          const baselineZips = await globby(".serverless/*.zip", { cwd });
+          await Promise.all(baselineZips.map((zipFile) => fs.copy(
+            path.join(cwd, zipFile),
+            path.join(baselineArchive, path.basename(zipFile))
+          )));
+        }
 
-        // Data.
-        // eslint-disable-next-line no-magic-numbers
-        const pct = ((pluginTime - baselineTime) / baselineTime * 100).toFixed(2);
-        const pluginRow = [scenario, mode, "jetpack", pluginTime, `**${pct} %**`];
+        // Report.
+        let pct;
+        if (baselineTime) {
+          // eslint-disable-next-line no-magic-numbers
+          pct = ((pluginTime - baselineTime) / baselineTime * 100).toFixed(2);
+        }
+        const pluginRow = [scenario, mode, "jetpack", pluginTime, pct ? `**${pct} %**` : ""];
 
         const resultsKey = `${scenario}/${mode}`;
         results[resultsKey] = (results[resultsKey] || []).concat([
           pluginRow,
-          [scenario, mode, "baseline", baselineTime, ""]
-        ]);
+          baselineTime ? [scenario, mode, "baseline", baselineTime, ""] : null
+        ].filter(Boolean));
       });
     })
   );